+++ src/main/java/ru/myapp/MainWebInitializer.java
@@ -1,0 +1,40 @@
+package ru.myapp;
+
+import jakarta.servlet.ServletContext;
+import jakarta.servlet.ServletRegistration;
+import org.springframework.web.WebApplicationInitializer;
+import org.springframework.web.context.ContextLoaderListener;
+import org.springframework.web.context.support.AnnotationConfigWebApplicationContext;
+import org.springframework.web.servlet.DispatcherServlet;
+
+public class MainWebInitializer implements WebApplicationInitializer {
+
+    @Override
+    public void onStartup(ServletContext servletContext) {
+
+//        // Create the 'root' Spring application context
+//        AnnotationConfigWebApplicationContext rootContext =
+//                new AnnotationConfigWebApplicationContext();
+//        rootContext.register(RootConfig.class);
+//
+//        // Manage the lifecycle of the root application context
+//        servletContext.addListener(new ContextLoaderListener(rootContext));
+//
+//        // Create the dispatcher servlet's Spring application context
+//        AnnotationConfigWebApplicationContext dispatcherContext =
+//                new AnnotationConfigWebApplicationContext();
+//        dispatcherContext.register(DispatcherConfig.class);
+//        dispatcherContext.setServletContext(servletContext);
+//        // Register and map the dispatcher servlet
+//        ServletRegistration.Dynamic dispatcher =
+//                servletContext.addServlet("dispatcher", new DispatcherServlet(dispatcherContext));
+//        dispatcher.setLoadOnStartup(1);
+//        dispatcher.addMapping("/");
+        AnnotationConfigWebApplicationContext context = new AnnotationConfigWebApplicationContext();
+        context.scan("ru.myapp");
+        servletContext.addListener(new ContextLoaderListener(context));
+        ServletRegistration.Dynamic servlet = servletContext.addServlet("spring mvc", new DispatcherServlet(context));
+        servlet.setLoadOnStartup(1);
+        servlet.addMapping("/");
+    }
+}
