+++ src/main/java/ru/myapp/dto/GroupRequestDto.java
@@ -1,0 +1,5 @@
+package ru.myapp.dto;
+
+public record GroupRequestDto(String name,
+                              String description) {
+}
