+++ src/main/java/ru/myapp/dto/GroupResponseDto.java
@@ -1,0 +1,9 @@
+package ru.myapp.dto;
+
+import java.util.Set;
+
+public record GroupResponseDto(String groupId,
+                               String name,
+                               String description,
+                               Set<UserResponseDtoShort> users) {
+}
