+++ src/main/java/ru/myapp/dto/GroupResponseDtoShort.java
@@ -1,0 +1,6 @@
+package ru.myapp.dto;
+
+public record GroupResponseDtoShort(String groupId,
+                                    String name,
+                                    String description) {
+}
