+++ src/main/java/ru/myapp/dto/UserRequestDto.java
@@ -1,0 +1,5 @@
+package ru.myapp.dto;
+
+public record UserRequestDto(String firstName,
+                             String lastName) {
+}
