+++ src/main/java/ru/myapp/dto/UserResponseDto.java
@@ -1,0 +1,9 @@
+package ru.myapp.dto;
+
+import java.util.Set;
+
+public record UserResponseDto(String userId,
+                              String firstName,
+                              String lastName,
+                              Set<GroupResponseDtoShort> groups) {
+}
