+++ src/main/java/ru/myapp/dto/UserResponseDtoShort.java
@@ -1,0 +1,6 @@
+package ru.myapp.dto;
+
+public record UserResponseDtoShort(String userId,
+                                   String firstName,
+                                   String lastName) {
+}
