+++ src/main/java/ru/myapp/error/ApiError.java
@@ -1,0 +1,7 @@
+package ru.myapp.error;
+
+public record ApiError(String error,
+                       String message,
+                       String status,
+                       String timestamp) {
+}
