+++ src/main/java/ru/myapp/repository/UserRepository.java
@@ -1,0 +1,15 @@
+package ru.myapp.repository;
+
+import org.springframework.data.jpa.repository.EntityGraph;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.lang.NonNull;
+import ru.myapp.model.User;
+
+import java.util.Optional;
+
+public interface UserRepository extends JpaRepository<User, Integer> {
+
+    @Override
+    @EntityGraph(attributePaths = "groups")
+    Optional<User> findById(@NonNull Integer userId);
+}
