--- src/main/java/spring/config/DatabaseInitializer.java
@@ -1,33 +1,0 @@
-package spring.config;
-
-import jakarta.annotation.PostConstruct;
-import org.springframework.core.io.ClassPathResource;
-import org.springframework.core.io.Resource;
-import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.jdbc.datasource.init.ScriptUtils;
-import org.springframework.stereotype.Component;
-
-import java.sql.SQLException;
-import java.util.Objects;
-
-@Component
-public class DatabaseInitializer {
-
-    private final JdbcTemplate jdbcTemplate;
-
-    public DatabaseInitializer(JdbcTemplate jdbcTemplate) {
-        this.jdbcTemplate = jdbcTemplate;
-    }
-
-    @PostConstruct
-    public void initializeDatabase() throws SQLException {
-        Resource schema = new ClassPathResource("schema.sql");
-        Resource data = new ClassPathResource("data.sql");
-        executeScript(schema);
-        executeScript(data);
-    }
-
-    private void executeScript(Resource resource) throws SQLException {
-        ScriptUtils.executeSqlScript(Objects.requireNonNull(jdbcTemplate.getDataSource()).getConnection(), resource);
-    }
-}
