--- src/main/java/spring/controllers/IndexController.java
@@ -1,15 +1,0 @@
-package spring.controllers;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-
-@Controller
-@RequestMapping("/")
-public class IndexController {
-
-    @GetMapping
-    public String welcome() {
-        return "welcome";
-    }
-}
