--- src/main/java/spring/dto/GroupRequestDto.java
@@ -1,4 +1,0 @@
-package spring.dto;
-
-public record GroupRequestDto(String name, String description) {
-}
