--- src/main/java/spring/dto/GroupResponseDto.java
@@ -1,4 +1,0 @@
-package spring.dto;
-
-public record GroupResponseDto(String groupId, String name, String description) {
-}
