--- src/main/java/spring/dto/UserRequestDto.java
@@ -1,4 +1,0 @@
-package spring.dto;
-
-public record UserRequestDto(String firstName, String lastName) {
-}
