--- src/main/java/spring/dto/UserResponseDto.java
@@ -1,4 +1,0 @@
-package spring.dto;
-
-public record UserResponseDto(String userId, String firstName, String lastName) {
-}
