--- src/main/java/spring/error/ApiError.java
@@ -1,7 +1,0 @@
-package spring.error;
-
-public record ApiError(String error,
-                       String message,
-                       String status,
-                       String timestamp) {
-}
