--- src/main/java/spring/error/ErrorHandler.java
@@ -1,34 +1,0 @@
-package spring.error;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.ControllerAdvice;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.ResponseBody;
-import org.springframework.web.bind.annotation.ResponseStatus;
-
-import java.time.LocalDateTime;
-
-import static utils.Constants.DATE_TIME_FORMATTER;
-
-@ControllerAdvice
-@ResponseBody
-public class ErrorHandler {
-
-    @ExceptionHandler
-    @ResponseStatus(HttpStatus.INTERNAL_SERVER_ERROR)
-    public ApiError handleThrowable(final Throwable e) {
-        return new ApiError("Internal Server error.",
-                e.getMessage(),
-                HttpStatus.INTERNAL_SERVER_ERROR.toString(),
-                LocalDateTime.now().format(DATE_TIME_FORMATTER));
-    }
-
-    @ExceptionHandler
-    @ResponseStatus(HttpStatus.NOT_FOUND)
-    public ApiError handleNotFoundException(final NotFoundException e) {
-        return new ApiError("The required resource was not found.",
-                e.getMessage(),
-                HttpStatus.NOT_FOUND.toString(),
-                LocalDateTime.now().format(DATE_TIME_FORMATTER));
-    }
-}
