--- src/main/java/spring/repository/UserRepository.java
@@ -1,9 +1,0 @@
-package spring.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-import spring.model.User;
-
-@Repository
-public interface UserRepository extends JpaRepository<User, Integer> {
-}
