+++ src/main/java/utils/Constants.java
@@ -4,6 +4,10 @@ import java.time.format.DateTimeFormatter;
 
 public class Constants {
 
+    public Constants() {
+        throw new RuntimeException();
+    }
+
     public static final String DATE_TIME_PATTERN = "yyyy-MM-dd HH:mm:ss";
 
     public static final DateTimeFormatter DATE_TIME_FORMATTER = DateTimeFormatter.ofPattern(DATE_TIME_PATTERN);
