+++ src/main/webapp/WEB-INF/view/index.jsp
@@ -1,0 +1,15 @@
+<%@ page language="java" contentType="text/html; charset=US-ASCII"
+         pageEncoding="US-ASCII" %>
+<!DOCTYPE html PUBLIC "-//W3C//DTD HTML 4.01 Transitional//EN" "https://www.w3.org/TR/html4/loose.dtd">
+<html>
+<head>
+    <meta http-equiv="Content-Type" content="text/html; charset=US-ASCII">
+    <title>First JSP PAGE</title>
+</head>
+<%@ page import="java.util.Date" %>
+<body>
+<h3>WELCOME TO MY APP</h3><br>
+<strong>Current Time is</strong>: <%=new Date() %>
+
+</body>
+</html>
